Fix underreporting of New Solar Power
New Solar Power is miscalculated when the main inverter is able to
charge the battery while still outputting to the rest of the house as
it was assumed that all of the PV output would go to the house. This
meant that any power going to the battery was missing from the
calculation. This adds that component back in.